+++ src/App.js
@@ -1,10 +1,10 @@
 import React from "react";
 import "./App.css";
+import { Switch, Route } from "react-router-dom";
 import Navbar from "./components/navbar";
 import Footer from "./components/footer";
 import Home from "./components/home";
 import About from "./components/about";
-
 
 function App() {
   return (
@@ -13,7 +13,10 @@ function App() {
         <Navbar></Navbar>
       </header>
       <main className='container-fluid flex-fill'>
-        <About />
+        <Switch>
+          <Route path='/about' component={About}></Route>
+          <Route path='/' exact component={Home}></Route>
+        </Switch>
       </main>
       <footer>
         <Footer />